1) Modified DoubleAuctionOrderBook to use TreeSets, LimitOrders to implement Comparable. This allowed to drop some third of code in OrderBook and obtain some speedup by removing Vector overhead; 2) DoubleAuctionOrderBook now implements cleanup where expired orders are removed and, returns per tick computed and negative spreads dealt with; 3) FarmerPlayer randomly drops LimitOrders and MarkerOrders to test order book scheme. Need to move from uniform to Poisson distributions to get to Farmer's specification; 4) Added histogram-based visualization for order book shape.
TO DO: I got completely lost what is ask, bid, sell, buy. There must be plenty of bugs on this level.

TO DO: Finish feedbacks from Market to Players on status of their orders.

TO DO: There is many ways of determining exact exchange prices and defining returns (end of period? integrated over tick?). Need to set down on one.

+++ src/model/market/books/DoubleAuctionOrderBook.java
@@ -127,8 +127,8 @@ public class DoubleAuctionOrderBook implements OrderBook {
 	// Returns total price of purchasing 'quantity' units if successful.
 	// Otherwise, throws LiquidityException, which contains the number
 	// successfully executed.
-	public synchronized double executeMarketOrder(OrderType type, int quantity) throws LiquidityException {
-		int origQuant = quantity;
+	public synchronized double executeMarketOrder(OrderType type, double quantity) throws LiquidityException {
+		int origQuant = (int) quantity;
 		double totalPrice = 0.0;
 		
 		Set<Map.Entry<Integer,Vector<LimitOrder>>> queueSet = null;
@@ -157,12 +157,12 @@ public class DoubleAuctionOrderBook implements OrderBook {
      			
      			if (o.quantityPending() > quantity) {
      				// this won't exhaust the limitOrder
-     				curQuantity = quantity;
+     				curQuantity = (int) quantity;
      				orders.add(0, o); // adds the order back in since
      			} else if (o.quantityPending() == quantity) {
-     				curQuantity = quantity;
+     				curQuantity = (int) quantity;
      			} else {
-     				curQuantity = quantity - o.quantityPending();
+     				curQuantity = (int) (quantity - o.quantityPending());
      			}
 
      			// Execute:
@@ -178,7 +178,7 @@ public class DoubleAuctionOrderBook implements OrderBook {
      		} 
      	}		
 		if (quantity != 0) {
-			throw new LiquidityException(origQuant - quantity, totalPrice); 
+			throw new LiquidityException((int) (origQuant - quantity), totalPrice); 
 		}
 
 		return totalPrice;
